backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
cors==1.0.1

# database and cache
SQLAlchemy==2.0.48
asyncpg==0.28.0
aioredis==2.0.1
aiosqlite==0.20.0
alembic==1.13.1

# auth
python-jose==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
Authlib==1.3.1
httpx==0.27.0
itsdangerous==2.2.0
